+++ CHANGES.rst
@@ -3,6 +3,7 @@ Changelog (nionutils)
 
 0.3.27 (UNRELEASED)
 -------------------
+- Add experiment stream, value change, and reactor functions.
 - Add method to sync but not close event loop.
 - Add a single item dispatcher to thread pool.